Fix specs and break out link subcomponents
