+++ src/components/link/external-link/fixtures.js
@@ -1,0 +1,9 @@
+export default {
+  default: {
+    children: "Hello World",
+    className: "",
+    target: undefined,
+    to: "/",
+    onClick: null
+  }
+}
