+++ src/components/link/external-link/index.js
@@ -1,0 +1,6 @@
+import component from "./component"
+import fixtures from "./fixtures"
+
+export default component
+
+export { component, fixtures }
