+++ src/components/link/helpers/index.js
@@ -1,0 +1,4 @@
+import isFile from "./is-file"
+import isGatsbyLink from "./is-gatsby-link"
+
+export { isFile, isGatsbyLink }
