+++ src/components/link/internal-link/fixtures.js
@@ -1,0 +1,10 @@
+export default {
+  default: {
+    activeClassName: "active",
+    activeStyle: {},
+    children: "Hello World",
+    className: "",
+    to: "/",
+    onClick: null
+  }
+}
