+++ src/components/link/internal-link/test.spec.js
@@ -1,0 +1,11 @@
+import React from "react"
+import renderer from "react-test-renderer"
+
+import { component as InternalLink, fixtures } from "."
+
+describe("InternalLink", () => {
+  it("renders correctly", () => {
+    const tree = renderer.create(<InternalLink {...fixtures.default} />).toJSON()
+    expect(tree).toMatchSnapshot()
+  })
+})
